at least it works ^_^

+++ engine/objectmanager.h
@@ -9,7 +9,7 @@ namespace D3 {
 
 struct ObjectManager{ // sizeof = 0xA10
     UCHAR       unknown_000[0x38];
-    ULONG       x038_Counter_CurrentFrame;
+    DWORD       x038_Counter_CurrentFrame;
     UCHAR       unknown_03c[0x958-0x3c];
     ExpandableContainer<Actor>* x958_RActors;
     UCHAR       unknown_95c[0x998-0x95c];
